Re-disabling one rowsum() test on Windows
In commit c7ee174 I enabled a rowsum() test that was previously disabled
on Windows (because it can't allocate a ~150 Mb vector when run in 32-bit
mode). I just wanted to give it a try. However, on today's Bioconductor
build report, the test is failing on Windows so I'm disabling it again on
this platform.